Add Zip support, change name of plugin to reflect extended functionality.
